+++ packages/opentelemetry-resources/src/platform/node/ServiceInstanceIDDetectorSync.ts
@@ -36,4 +36,5 @@ class ServiceInstanceIDDetectorSync implements DetectorSync {
 /**
  * @experimental
  */
-export const serviceInstanceIDDetectorSync = new ServiceInstanceIDDetectorSync();
+export const serviceInstanceIDDetectorSync =
+  new ServiceInstanceIDDetectorSync();