+++ CHANGELOG.md
@@ -12,7 +12,7 @@ For experimental package changes, see the [experimental CHANGELOG](experimental/
 ### :rocket: (Enhancement)
 
 * feat(sdk-trace-base): log resource attributes in ConsoleSpanExporter [#4605](https://github.com/open-telemetry/opentelemetry-js/pull/4605) @pichlermarc
-* feat(resources): new experimental detector ServiceInstanceIDDetector that sets the value for `service.instance.id` as random UUID.
+* feat(resources): new experimental detector ServiceInstanceIDDetectorSync that sets the value for `service.instance.id` as random UUID.
 
 ### :bug: (Bug Fix)